requirements.txt
Flask>=2.1.0
# SQLAlchemy>=1.4.22  # unsure if necessary or overdescribed
flask-sqlalchemy>=2.5.1

# If using mysql as in tutorial
pymysql
cryptography
mysql-connector-python  # for initial db creation